test: extract equation lineage family
